requirements.txt
PyJWT~=2.3.0
Django~=3.2.6
djangorestframework~=3.13.1
psycopg2~=2.9.3